# week7/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Codegen Demo API
  description: Product management API used to demo FastAPI backend code generation and database integration.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
tags:
  - name: Products
    description: CRUD APIs for products
paths:
  /products:
    get:
      tags:
        - Products
      summary: Get product list
      operationId: get_products
      parameters:
        - name: name
          in: query
          description: Partial-match name filter
          required: false
          schema:
            type: string
        - name: min_price
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            format: float
        - name: max_price
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Successfully returned list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: Create a new product
      operationId: create_product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product details by ID
      operationId: get_product_by_id
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
    put:
      tags:
        - Products
      summary: Update an entire product record
      operationId: update_product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '200':
          description: Record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
    delete:
      tags:
        - Products
      summary: Delete a product
      operationId: delete_product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted successfully (no content returned)
        '404':
          description: Product not found

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stock_quantity
      properties:
        id:
          type: integer
          description: Product ID
        name:
          type: string
          description: Product name
        price:
          type: number
          format: float
          description: Product price
        stock_quantity:
          type: integer
          description: Remaining stock quantity
    ProductCreate:
      type: object
      description: Schema used by POST and PUT (ID omitted because DB auto-increments it)
      required:
        - name
        - price
        - stock_quantity
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        stock_quantity:
          type: integer