# week7/prepared_data_module/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-04-16T09:43:20+00:00

from __future__ import annotations

from typing import List, Optional, Union

from fastapi import FastAPI, Path, Depends, HTTPException, status
from sqlalchemy.orm import Session

from .models import Product, ProductCreate
from .database import engine, get_db
from . import db_models, product_repo

# Initialize SQLite database tables
db_models.Base.metadata.create_all(bind=engine)

app = FastAPI(
    title='Swagger Codegen Demo API',
    description='Product management API used to demo FastAPI backend code generation and database integration.',
    version='1.0.0',
    servers=[
        {'url': 'http://localhost:8000', 'description': 'Local development server'}
    ],
)


@app.get('/products', response_model=List[Product], tags=['Products'])
def get_products(
    name: Optional[str] = None,
    min_price: Optional[float] = None,
    max_price: Optional[float] = None,
    db: Session = Depends(get_db)
) -> List[Product]:
    """
    Get product list.
    """
    return product_repo.get_products(db, name=name, min_price=min_price, max_price=max_price)


@app.post(
    '/products',
    response_model=Product,
    status_code=status.HTTP_201_CREATED,
    tags=['Products'],
)
def create_product(body: ProductCreate, db: Session = Depends(get_db)) -> Product:
    """
    Create a new product.
    """
    return product_repo.create_product(db, body)


@app.get('/products/{productId}', response_model=Product, tags=['Products'])
def get_product_by_id(product_id: int = Path(..., alias='productId'), db: Session = Depends(get_db)) -> Product:
    """
    Get product details by ID.
    """
    db_product = product_repo.get_product_by_id(db, product_id)
    if not db_product:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found")
    return db_product


@app.put('/products/{productId}', response_model=Product, tags=['Products'])
def update_product(
    product_id: int = Path(..., alias='productId'), body: ProductCreate = ..., db: Session = Depends(get_db)
) -> Product:
    """
    Update an entire product record.
    """
    db_product = product_repo.update_product(db, product_id, body)
    if not db_product:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found")
    return db_product


@app.delete('/products/{productId}', status_code=status.HTTP_204_NO_CONTENT, tags=['Products'])
def delete_product(product_id: int = Path(..., alias='productId'), db: Session = Depends(get_db)) -> None:
    """
    Delete a product.
    """
    success = product_repo.delete_product(db, product_id)
    if not success:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found")