# week7/server_app/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-04-16T09:43:20+00:00

from __future__ import annotations

from typing import List, Optional, Union

from fastapi import FastAPI, Path

from .models import Product, ProductCreate

app = FastAPI(
    title='Swagger Codegen Demo API',
    description='Product management API used to demo FastAPI backend code generation and database integration.',
    version='1.0.0',
    servers=[
        {'url': 'http://localhost:8000', 'description': 'Local development server'}
    ],
)


@app.get('/products', response_model=List[Product], tags=['Products'])
def get_products(
    name: Optional[str] = None,
    min_price: Optional[float] = None,
    max_price: Optional[float] = None,
) -> List[Product]:
    """
    Get product list.
    """
    pass


@app.post(
    '/products',
    response_model=None,
    responses={'201': {'model': Product}},
    tags=['Products'],
)
def create_product(body: ProductCreate) -> Optional[Product]:
    """
    Create a new product.
    """
    pass


@app.get('/products/{productId}', response_model=Product, tags=['Products'])
def get_product_by_id(product_id: int = Path(..., alias='productId')) -> Product:
    """
    Get product details by ID.
    """
    pass


@app.put('/products/{productId}', response_model=Product, tags=['Products'])
def update_product(
    product_id: int = Path(..., alias='productId'), body: ProductCreate = ...
) -> Product:
    """
    Update an entire product record.
    """
    pass


@app.delete('/products/{productId}', response_model=None, tags=['Products'])
def delete_product(product_id: int = Path(..., alias='productId')) -> None:
    """
    Delete a product.
    """
    pass