# week7/server_app/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-04-16T09:43:20+00:00

from __future__ import annotations

from pydantic import BaseModel, Field


class Product(BaseModel):
    id: int = Field(..., description='Product ID')
    name: str = Field(..., description='Product name')
    price: float = Field(..., description='Product price')
    stock_quantity: int = Field(..., description='Remaining stock quantity')


class ProductCreate(BaseModel):
    name: str
    price: float
    stock_quantity: int